refactor(vehicle_info): rename utils to vehicle_utils
