+++ vehicle_info/vehicle_utils.py
